Conflicts:
	res/pieces/pieces-biz/src/main/webapp/assets/js/common.js
	res/pieces/pieces-biz/src/main/webapp/assets/js/layer/skin/default/loading-0.gif

+++ res/pieces/pieces-boss/src/main/java/com/pieces/boss/controller/CategoryController.java
@@ -290,10 +290,11 @@ public class CategoryController {
 							  HttpServletResponse response,
 							  @PathVariable("id") Integer id,
 							  ModelMap model){
-		CommodityVO vo = commodityService.findCommodityByBreedId(id);
-		if(vo != null){
+		List<CommodityVO> vos = commodityService.findCommodityByBreedId(id);
+		if(ValidUtils.listNotBlank(vos)){
 			Result result = new Result(false).info("该品种下已有商品，请先将所有商品移除后再删除。");
 	        WebUtil.printJson(response, result);
+	        return;
 		}
 		categoryService.deleteById(id);
 		Result result = new Result(true);
